working code

+++ run.py
@@ -1,18 +1,12 @@
 #!/usr/bin/python3
 
-#TODO: convert Activity type :)
-
 import os
-import os.path
 import re
 import subprocess
-import glob
 import csv
-import collections
-
-# CONSTANTS ------------------------------------
 import sys
 
+# CONSTANTS ------------------------------------
 UNZIPPED_FOLDER = "unzipped"
 STRAVA_ACTIVITIES_SUBFOLDER = "activities"
 STRAVA_ACTIVITIES_FILE = "activities.csv"
@@ -22,7 +16,6 @@ STRAVA_ACTIVITIES_FILE = "activities.csv"
 def unzip(file):
     import gzip
     import shutil
-    print(file)
     with gzip.open(file, 'rb') as f_in:
         with open(file.lower().replace('.gz', ''), 'wb') as f_out:
             shutil.copyfileobj(f_in, f_out)
@@ -46,10 +39,11 @@ def main():
     global STRAVA_ACTIVITIES_SUBFOLDER
     global STRAVA_ACTIVITIES_FILE
 
+
     strava_unzipped_folder = os.path.join(".", UNZIPPED_FOLDER)
     strava_unzipped_activities_folder = os.path.join(".", UNZIPPED_FOLDER, STRAVA_ACTIVITIES_SUBFOLDER)
     strava_unzipped_activities_file = os.path.join(".", UNZIPPED_FOLDER, STRAVA_ACTIVITIES_FILE)
-    gpsbabel_command = "gpsbabel -w -t -i gtrnctr -f {filename_tcx} -o gpx,gpxver=1.1 -F {filename_gpx}"
+    gpsbabel_command = "gpsbabel -w -t -i gtrnctr -f {filename_tcx} -o gpx,gpxver=1.1 -F {filename_gpx} > /dev/null"
 
     #check ----------------
     if not os.path.exists(strava_unzipped_folder):
@@ -65,33 +59,40 @@ def main():
     try:
         with open(strava_unzipped_activities_file, newline='\n', encoding='utf-8') as csvfile:
             reader = csv.DictReader(csvfile)
-            cnt = 0
-            for row in reader:
-                cnt += 1
-            print("Found {cnt} activities ".format(cnt=str(cnt)))
-
+            activities_list = list(reader)
     except Exception as ex:
         die("Cannot read '{path}'. {err}".format(path=strava_unzipped_activities_file, err=str(ex)))
 
-    print("Unzipping ...");
-    #sys.stdout.write("Progress: ")
+    cnt = len(activities_list)
+    print("Found {cnt} activities ".format(cnt=str(cnt)))
+
+    print("Unzipping ...")
+    sys.stdout.write("Progress: ")
     ucnt = 0
-    for gz in glob.glob(os.path.join(strava_unzipped_activities_folder, "*.gz")):
+    for line in activities_list:
+        x = 0
+        filename = os.path.join(strava_unzipped_folder, line['filename'])
         ucnt += 1
-        #sys.stdout.write("{:2.0%}".format(ucnt / cnt))
-        #sys.stdout.flush()
-        unzip(os.path.join(strava_unzipped_activities_folder, gz))
-        #sys.stdout.write("\b\b\b\b")
-        #sys.stdout.flush()
-    #sys.stdout.write("\n")
+        sys.stdout.write("{:2.0%}".format(ucnt / cnt).rjust(5, " "))
+        sys.stdout.flush()
+        if filename.endswith(".gz") and os.path.exists(filename):
+            unzip(filename)
+        line['filename'] = line['filename'].replace(".gz", "")
+        sys.stdout.write("\b\b\b\b\b")
+        sys.stdout.flush()
+        x += 1
+    sys.stdout.write("\n")
 
-    for root, dirs, files in os.walk(strava_unzipped_activities_folder):
-        for filename in files:
-            extension = os.path.splitext(filename)[1][1:].strip().lower()
+    print("Normalizing ...")
+    for line in activities_list:
+        filename = os.path.join(strava_unzipped_folder, line['filename'])
+        extension = os.path.splitext(filename)[1][1:].strip().lower()
+        # strip spaces from gpx
+        print(filename.ljust(40," ") + "| " + line["type"].ljust(13," ") + "| " + str(line["name"])[:70])
 
-            # strip spaces from gpx
+        if os.path.exists(filename):
             if extension == "gpx":
-                with open(os.path.join(strava_unzipped_activities_folder, filename), encoding='utf-8') as file:
+                with open(filename, encoding='utf-8') as file:
                     content = file.read()
                     fixed = re.sub(r'>\s\s+<', '><', content).strip()
                     with open(filename, "w", encoding='utf-8') as filew:
@@ -99,21 +100,47 @@ def main():
 
             # strip spaces from tcx
             if extension == "tcx":
-                with open(os.path.join(strava_unzipped_activities_folder, filename), encoding='utf-8') as file:
+                with open(filename, encoding='utf-8') as file:
                     content = file.read()
                     fixed = re.sub(r'>\s\s+<', '><', content).strip()
                     with open(filename, "w", encoding='utf-8') as filew:
                         filew.write(fixed)
                         print(re.findall(r'<[Aa][Cc][Tt][Ii][Vv][Ii][Tt][Yy][ ]{1,}[Ss][Pp][Oo][Rr][Tt]="([^"]*)">', fixed))
-                        # TODO: save artivity type #read from .activitues!
 
                     filename_gpx = filename.replace(".tcx", ".gpx")
                     rc = exec_cmd(gpsbabel_command.format(
-                        filename_gpx=os.path.join(strava_unzipped_activities_folder, filename_gpx),
-                        filename_tcx=os.path.join(strava_unzipped_activities_folder, filename))
+                        filename_gpx=filename_gpx,
+                        filename_tcx=filename)
                     )
                     if not rc:
                         print('error')
+                    else:
+                        os.remove(filename)
+                        line['filename'] = line['filename'].replace(".tcx", ".gpx")
+
+    new_activities_file = strava_unzipped_activities_file + ".original"
+    print("Backing up the original activities as '{new}' ...".format(new=new_activities_file))
+    os.rename(strava_unzipped_activities_file, new_activities_file)
+
+    print("Saving new activities ...")
+    with open(strava_unzipped_activities_file, 'w', newline='\n', encoding='utf-8') as csvfile:
+
+        fieldnames = ['id',
+                      'date',
+                      'name',
+                      'type',
+                      'description',
+                      'elapsed_time',
+                      'distance',
+                      'commute',
+                      'gear',
+                      'filename'
+                     ]
+        writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+        writer.writeheader()
+
+        for line in activities_list:
+            writer.writerow(line)
 
 
 if __name__ == "__main__":
